fix(server): erro de timeout após atualizar perfil
Este erro ocorria de vez em quando logo após enviar uma requisição para atualizar o perfil do
usuário. Ele levava o servidor a não conseguir mais se conectar com o banco de dados. Era informado
que houve timeout  no Knex adquirir uma conexão com o banco e que a causa era de que o 'pull'
provavelmente estava cheio. Investiguei e o problema estava na tentativa de uma consulta INSERT sem
registros a inserir, o que causava o erro SQLITE_MISUSE, que foi erroneamente tratado. Quando recebi
este erro MISUSE pela primeira vez eu entendi que era causado por quando executava-se uma transação
em que seus UPDATES mantiam o mesmo valor anterior e pensava que o erro era disparado ao realizar
commit da transaction, levando a tratar o erro no catch sem realizar um rollback. No entanto, isso
estava errado. O erro MISUSE era disparado no supracitado INSERT sem registros. Agora, o INSERT
somente é realizado se houver registros a inserir. Com isso, o erro de TIMEOUT não aconteceu mais
nos meus testes.

+++ server/src/controllers/UsersController.ts
@@ -206,9 +206,12 @@ export class UsersController {
           })
           .map((myScheduleToDelete) => myScheduleToDelete.id);
 
-        await trx('class_schedule')
-          .whereIn('id', classSchedulesToDeleteIds)
-          .delete();
+        if (classSchedulesToDeleteIds.length > 0) {
+          await trx('class_schedule')
+            .whereIn('id', classSchedulesToDeleteIds)
+            .delete();
+        }
+
         // eslint-disable-next-line no-restricted-syntax
         for (const schedule of classSchedulesToUpdate) {
           const formattedScheduleToInsert: ClassScheduleToUpdateDTO = {
@@ -232,22 +235,25 @@ export class UsersController {
           });
         });
       }
-      await trx('class_schedule').insert(classSchedulesToInsert);
+
+      if (classSchedulesToInsert.length > 0) {
+        await trx('class_schedule').insert(classSchedulesToInsert);
+      }
 
       await trx.commit();
 
       response.status(200).json();
     } catch (error: any) {
-      if (error?.code === 'SQLITE_MISUSE') {
-        response.status(200).json();
-      } else {
-        console.error(error);
-        await trx.rollback();
+      console.error({
+        error,
+        code: error?.code,
+        message: error?.message,
+      });
+      await trx.rollback();
 
-        response.status(400).json({
-          error: 'Erro inesperado ao atualizar perfil',
-        });
-      }
+      response.status(400).json({
+        error: 'Erro inesperado ao atualizar perfil',
+      });
     }
   }
